Refactor the log_chef method to accept a block and lazy load
Unlike the stdlib Logger class the Chef::Log class does not verify that
the severity is greater than or equal to the the current log level
before yielding to the given block. This results in **abysmal**
performance with large log messages due to the time taken to construct
the String even if we do not need it.

Refactoring the method to immediately return if the severity is less
than the current level results in a ~2000% run time reduction from (10-15
minutes to 30 seconds) for a reasonable size test set.